migrate make_config
