+++ src/hydra_zen/structured_configs/_make_config.py
@@ -1,39 +1,19 @@
 # Copyright (c) 2023 Massachusetts Institute of Technology
 # SPDX-License-Identifier: MIT
-import warnings
-from collections import Counter
-from dataclasses import (  # use this for runtime checks
-    InitVar,
-    dataclass,
-    make_dataclass,
-)
-from typing import Any, Dict, List, Optional, Tuple, Type, Union, cast
+from typing import Optional, Tuple, Type, Union
 
 from typing_extensions import Literal
 
-from hydra_zen.errors import HydraZenDeprecationWarning
-from hydra_zen.structured_configs import _utils
-from hydra_zen.structured_configs._implementations import _BUILDS_CONVERT_SETTINGS
 from hydra_zen.typing import DataclassOptions, SupportedPrimitive
 from hydra_zen.typing._implementations import (
     AllConvert,
     DataClass,
     DataClass_,
     DefaultsList,
-    Field,
     ZenConvert,
 )
 
-from ._globals import (
-    CONVERT_FIELD_NAME,
-    DEFAULTS_LIST_FIELD_NAME,
-    PARTIAL_FIELD_NAME,
-    RECURSIVE_FIELD_NAME,
-    ZEN_PARTIAL_FIELD_NAME,
-    ZEN_TARGET_FIELD_NAME,
-)
-from ._implementations import _retain_type_info, builds
-from ._type_guards import safe_getattr, uses_zen_processing
+from ._implementations import ZenField, builds
 
 __all__ = ["ZenField", "make_config"]
 
@@ -41,63 +21,6 @@ __all__ = ["ZenField", "make_config"]
 class NOTHING:
     def __init__(self) -> None:
         raise TypeError("`NOTHING` cannot be instantiated")
-
-
-@dataclass(unsafe_hash=True)
-class ZenField:
-    """
-    ZenField(hint=Any, default=<class 'NOTHING'>, name=<class 'NOTHING'>)
-
-    Specifies a field's name and/or type-annotation and/or default value.
-    Designed to specify fields in `make_config`.
-
-    See the Examples section of the docstring for `make_config` for examples of using
-    `ZenField`.
-
-    Parameters
-    ----------
-    hint : type, optional (default=Any)
-    default : Any, optional
-    name : str, optional
-
-    Notes
-    -----
-    ``default`` will be returned as an instance of :class:`dataclasses.Field`.
-    Mutable values (e.g. lists or dictionaries) passed to ``default`` will automatically
-    be "packaged" in a default-factory function [1]_.
-
-    A type passed to ``hint`` will automatically be "broadened" such that the resulting
-    type is compatible with Hydra's set of supported type annotations [2]_.
-
-    References
-    ----------
-    .. [1] https://docs.python.org/3/library/dataclasses.html#default-factory-functions
-    .. [2] https://hydra.cc/docs/next/tutorials/structured_config/intro/#structured-configs-supports
-
-    See Also
-    --------
-    make_config: create a config with customized field names, default values, and annotations.
-    """
-
-    hint: type = Any
-    default: Union[SupportedPrimitive, Field[Any]] = _utils.field(default=NOTHING)
-    name: Union[str, Type[NOTHING]] = NOTHING
-    zen_convert: InitVar[Optional[ZenConvert]] = None
-
-    def __post_init__(self, zen_convert: Optional[ZenConvert]) -> None:
-        if not isinstance(self.name, str):
-            if self.name is not NOTHING:
-                raise TypeError(f"`ZenField.name` expects a string, got: {self.name}")
-        convert_settings = _utils.merge_settings(zen_convert, _BUILDS_CONVERT_SETTINGS)
-        del zen_convert
-
-        self.hint = builds._sanitized_type(self.hint)
-
-        if self.default is not NOTHING:
-            self.default = builds._sanitized_field(
-                self.default,
-                convert_dataclass=convert_settings["dataclass"],
-            )
 
 
 _MAKE_CONFIG_SETTINGS = AllConvert(dataclass=False, flat_target=False)
@@ -349,201 +272,7 @@ def make_config(
 
     See :ref:`data-val` for more general data validation capabilities via hydra-zen.
     """
-    convert_settings = _utils.merge_settings(zen_convert, _MAKE_CONFIG_SETTINGS)
-    convert_settings = cast(ZenConvert, convert_settings)
-    del zen_convert
-
-    if zen_dataclass is None:
-        zen_dataclass = {}
-
-    # initial validation
-    _utils.parse_dataclass_options(zen_dataclass)
-
-    if "frozen" in fields_as_kwargs:
-        warnings.warn(
-            HydraZenDeprecationWarning(
-                "Specifying `builds(frozen=<...>)` is deprecated. Instead, "
-                "specify `builds(zen_dataclass={'frozen': <...>})"
-            ),
-            stacklevel=2,
-        )
-        zen_dataclass["frozen"] = fields_as_kwargs.pop("frozen")  # type: ignore
-
-    if "config_name" in fields_as_kwargs:
-        warnings.warn(
-            HydraZenDeprecationWarning(
-                "Specifying `make_config(config_name=<...>)` is deprecated. "
-                "Instead specify `make_config(zen_dataclass={'cls_name': <...>})"
-            ),
-            stacklevel=2,
-        )
-        zen_dataclass["cls_name"] = fields_as_kwargs.pop("config_name")  # type: ignore
-
-    if not bases:
-        bases = zen_dataclass.get("bases", ())
-
-    zen_dataclass.setdefault("cls_name", "Config")
-    dataclass_options = _utils.parse_dataclass_options(zen_dataclass)
-
-    for _field in fields_as_args:
-        if not isinstance(_field, (str, ZenField)):
-            raise TypeError(
-                f"`fields_as_args` can only consist of field-names (i.e. strings) or "
-                f"`ZenField` instances. Got: "
-                f"{', '.join(str(x) for x in fields_as_args if not isinstance(x, (str, ZenField)))}"
-            )
-        if isinstance(_field, ZenField) and _field.name is NOTHING:
-            raise ValueError(
-                f"All `ZenField` instances specified in `fields_as_args` must have a "
-                f"name associated with it. Got: {_field}"
-            )
-    for name, _field in fields_as_kwargs.items():
-        if isinstance(_field, ZenField):
-            if _field.name is not NOTHING and _field.name != name:
-                raise ValueError(
-                    f"`fields_as_kwargs` specifies conflicting names: the kwarg {name} "
-                    f"is associated with a `ZenField` with name {_field.name}"
-                )
-            else:
-                _field.name = name
-
-    if fields_as_args:
-        all_names = [f.name if isinstance(f, ZenField) else f for f in fields_as_args]
-        all_names.extend(fields_as_kwargs)
-
-        if len(all_names) != len(set(all_names)):
-            raise ValueError(
-                f"`fields_as_args` cannot specify the same field-name multiple times."
-                f" Got multiple entries for:"
-                f" {', '.join(str(n) for n, count in Counter(all_names).items() if count > 1)}"
-            )
-        for _name in all_names:
-            if isinstance(_name, str) and _name.startswith("_zen_"):
-                raise ValueError(
-                    f"The field-name specified via `{_name}=<...>` is reserved by hydra-zen."
-                    " You can manually create a dataclass to utilize this name in a structured config."
-                )
-        del all_names
-
-    if "defaults" in fields_as_kwargs:
-        if hydra_defaults is not None:
-            raise TypeError(
-                "`defaults` and `hydra_defaults` cannot be specified simultaneously"
-            )
-        _defaults = fields_as_kwargs.pop("defaults")
-
-        if not isinstance(_defaults, ZenField):  # pragma: no branch
-            hydra_defaults = _defaults  # type: ignore
-
-    # validate hydra-args via `builds`
-    # also check for use of reserved names
-    _tmp: Any = None
-
-    builds(
-        dict,
-        hydra_convert=hydra_convert,
-        hydra_recursive=hydra_recursive,
-        hydra_defaults=hydra_defaults,
-        **{k: _tmp for k in fields_as_kwargs},
-    )
-
-    normalized_fields: Dict[str, ZenField] = {}
-
-    for _field in fields_as_args:
-        if isinstance(_field, str):
-            normalized_fields[_field] = ZenField(
-                name=_field, hint=Any, zen_convert=convert_settings
-            )
-        else:
-            assert isinstance(_field.name, str)
-            normalized_fields[_field.name] = _field
-
-    for name, value in fields_as_kwargs.items():
-        if not isinstance(value, ZenField):
-            normalized_fields[name] = ZenField(
-                name=name,
-                default=value,
-                zen_convert=convert_settings,
-            )
-        else:
-            normalized_fields[name] = value
-
-    # fields without defaults must come first
-    config_fields: List[Union[Tuple[str, type], Tuple[str, type, Any]]] = [
-        (str(f.name), f.hint)
-        for f in normalized_fields.values()
-        if f.default is NOTHING
-    ]
-
-    config_fields.extend(
-        [
-            (
-                str(f.name),
-                (
-                    # f.default: Field
-                    # f.default.default: Any
-                    f.hint
-                    if _retain_type_info(
-                        type_=f.hint,
-                        value=f.default.default,  # type: ignore
-                        hydra_recursive=hydra_recursive,
-                    )
-                    else Any
-                ),
-                f.default,
-            )
-            for f in normalized_fields.values()
-            if f.default is not NOTHING
-        ]
-    )
-
-    if hydra_recursive is not None:
-        config_fields.append(
-            (
-                RECURSIVE_FIELD_NAME,
-                bool,
-                _utils.field(default=hydra_recursive, init=False),
-            )
-        )
-
-    if hydra_convert is not None:
-        config_fields.append(
-            (CONVERT_FIELD_NAME, str, _utils.field(default=hydra_convert, init=False))
-        )
-
-    if hydra_defaults is not None:
-        hydra_defaults = builds._sanitize_collection(
-            hydra_defaults, convert_dataclass=False
-        )
-        config_fields.append(
-            (
-                DEFAULTS_LIST_FIELD_NAME,
-                List[Any],
-                _utils.field(default_factory=lambda: list(hydra_defaults), init=False),
-            )
-        )
-
-    dataclass_options["bases"] = bases
-    module = dataclass_options.pop("module", None)
-    assert _utils.parse_strict_dataclass_options(dataclass_options), dataclass_options
-
-    out = make_dataclass(fields=config_fields, **dataclass_options)
-
-    if module is not None:
-        out.__module__ = module
-
-    if hasattr(out, ZEN_TARGET_FIELD_NAME) and not uses_zen_processing(out):
-        raise ValueError(
-            f"{out.__name__} inherits from base classes that overwrite some fields "
-            f"associated with zen-processing features. As a result, this config will "
-            f"not instantiate correctly."
-        )
-    if safe_getattr(out, PARTIAL_FIELD_NAME, False) and uses_zen_processing(out):
-        raise ValueError(
-            f"{out.__name__} specifies both `{PARTIAL_FIELD_NAME}=True` and `"
-            f"{ZEN_PARTIAL_FIELD_NAME}=True`. This config will not instantiate "
-            f"correctly. This is typically caused by inheriting from multiple, "
-            f"conflicting configs."
-        )
-
-    return cast(Type[DataClass], out)
+    _locals = locals().copy()
+    fields_as_args = _locals.pop("fields_as_args")
+    fields_as_kwargs = _locals.pop("fields_as_kwargs")
+    return builds.make_config(*fields_as_args, **_locals, **fields_as_kwargs)  # type: ignore
